+++ src/rider/main/kotlin/com/github/rafaelldi/diagnosticsclientplugin/actions/memoryDump/CollectMemoryDumpAction.kt
@@ -1,11 +1,11 @@
-package com.github.rafaelldi.diagnosticsclientplugin.actions
+package com.github.rafaelldi.diagnosticsclientplugin.actions.memoryDump
 
 import com.github.rafaelldi.diagnosticsclientplugin.dialogs.MemoryDumpDialog
 import com.github.rafaelldi.diagnosticsclientplugin.services.MemoryDumpController
 import com.github.rafaelldi.diagnosticsclientplugin.services.MemoryDumpSettings
 import com.github.rafaelldi.diagnosticsclientplugin.toolWindow.components.LocalProcessNode
+import com.github.rafaelldi.diagnosticsclientplugin.toolWindow.tabs.ChartProtocolSessionTab
 import com.github.rafaelldi.diagnosticsclientplugin.toolWindow.tabs.ProcessExplorerTab
-import com.github.rafaelldi.diagnosticsclientplugin.utils.toDotNetProcess
 import com.intellij.openapi.actionSystem.ActionUpdateThread
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
@@ -15,13 +15,8 @@ import com.jetbrains.rd.platform.util.lifetime
 class CollectMemoryDumpAction : AnAction() {
     override fun actionPerformed(event: AnActionEvent) {
         val project = event.project ?: return
-        val tree = event.getData(ProcessExplorerTab.PROCESS_TREE) ?: return
-
-        val selectedProcessNode = tree.selectedNode as? LocalProcessNode ?: return
-        val selectedProcess = selectedProcessNode.toDotNetProcess()
-        val processes = tree.getLocalProcessNodes().map { it.toDotNetProcess() }
-
-        val dialog = MemoryDumpDialog(project, selectedProcess, processes)
+        val pid = getProcessId(event) ?: return
+        val dialog = MemoryDumpDialog(pid, project)
         if (dialog.showAndGet()) {
             val model = dialog.getModel()
             MemoryDumpSettings.getInstance(project).update(model)
@@ -33,19 +28,23 @@ class CollectMemoryDumpAction : AnAction() {
 
     override fun update(event: AnActionEvent) {
         val project = event.project
-        val tree = event.getData(ProcessExplorerTab.PROCESS_TREE)
-        if (project == null || tree == null) {
-            event.presentation.isEnabledAndVisible = false
-            return
-        }
-
-        val processNode = tree.selectedNode as? LocalProcessNode
-        if (processNode == null) {
+        val pid = getProcessId(event)
+        if (project == null || pid == null) {
             event.presentation.isEnabledAndVisible = false
             return
         }
 
         event.presentation.isEnabledAndVisible = true
+    }
+
+    private fun getProcessId(event: AnActionEvent): Int? {
+        val sessionPid = event.getData(ChartProtocolSessionTab.SESSION_PROCESS_ID)
+        if (sessionPid != null) return sessionPid
+
+        val tree = event.getData(ProcessExplorerTab.PROCESS_TREE)
+        if (tree != null) return (tree.selectedNode as? LocalProcessNode)?.processId
+
+        return null
     }
 
     override fun getActionUpdateThread(): ActionUpdateThread = ActionUpdateThread.BGT